Project Creation.
Creating a React Project. Adding images to the project for the site. Creating a folder of site components. Adding the order of components to the "App.js" file.  Creating the "Data.js" file, in which we create arrays to use the information on the site. Adding the file "index.css" where the styles for Tailwind are written. Also added the file "workoutSlider" where the styles for the slider we will use are written.